Move existing package to at_app directory
