--- .metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: f4abaa0735eba4dfd8f33f73363911d63931fe03
-  channel: stable
-
-project_type: package
