+++ at_app/example/example.md
@@ -1,0 +1,30 @@
+# Examples
+
+## Installing the executable
+
+```sh
+flutter pub global activate at_app
+```
+
+## Creating an @platform app
+
+```sh
+at_app create path/to/output/directory
+```
+
+Windows Users:
+
+```sh
+at_app.bat create path/to/output/directory
+```
+
+## To See Full Usage Options
+
+```sh
+at_app create -h
+```
+
+Windows Users:
+```sh
+at_app.bat create -h
+```
