+++ at_app/lib/at_app.dart
@@ -1,0 +1,3 @@
+library at_app;
+
+export 'src/command_runner.dart';
