+++ at_app/lib/src/models/command_status.dart
