+++ at_app/lib/src/models/exceptions/cache_package_exception.dart
