+++ at_app/lib/src/models/exceptions/flutter_exception.dart
