+++ at_app/lib/src/models/exceptions/template_exception.dart
