+++ at_app/lib/src/models/file_manager.dart
