+++ at_app/lib/src/models/template_file_manager_base.dart
