+++ at_app/lib/src/services/template_generator.dart
