+++ at_app/lib/src/services/template_generator_services/config_manager_services/gitignore_manager.dart
