+++ at_app/lib/src/services/template_generator_services/config_manager_services/gradle_properties_manager.dart
