--- bin/at_app.dart
@@ -1,5 +1,0 @@
-import 'package:at_app/at_app.dart';
-
-Future<void> main(List<String> args) async {
-  await AtCommandRunner().run(args);
-}
