--- lib/at_app.dart
@@ -1,3 +1,0 @@
-library at_app;
-
-export 'src/command_runner.dart';
